🧪 [testing improvement] Separate and expand use.local.desc tests
